finished release_dateprediction
