+++ scraping/release_date_model/filter_data.py
@@ -8,22 +8,12 @@ def get_data_for_model(input_file, output_file):
 
     data_for_model = []
     for sneaker in data:
-        release_date = sneaker.get('release_date')
-        if release_date is not None:
-            release_date = int(release_date)
-            year = release_date // 10000
-            month  = (release_date % 10000) // 100
-            day = release_date % 100
-        else:
-            year = None
-            month = None
-            day = None
         sneaker_dict = {
             'sneaker_name' : sneaker.get('sneaker_name'),
             'sneaker_color': sneaker.get('sneaker_color'),
-            'release_year' : year,
-            'release_month': month,
-            'release_day' : day,
+            'release_year' : sneaker.get('release_year'),
+            'release_month': sneaker.get('release_month'),
+            'release_day' : sneaker.get('release_day'),
         }
         data_for_model.append(sneaker_dict)
 
@@ -57,7 +47,14 @@ def get_data_to_predict(input_file, output_file):
     testing_data = []
     for sneaker in data:
         if sneaker.get('release_year') is None:
-            testing_data.append(sneaker)
+            sneaker_dict = {
+                'sneaker_name': sneaker.get('sneaker_name'),
+                'sneaker_color': sneaker.get('sneaker_color'),
+                'release_year': sneaker.get('release_year'),
+                'release_month': sneaker.get('release_month'),
+                'release_day': sneaker.get('release_day'),
+            }
+            testing_data.append(sneaker_dict)
 
     with open(output_file, 'w') as file:
         json.dump(testing_data, file, indent = 4)
